movves withProgressInWindow into separate folder
