+++ src/dependencies/GitHubZipExtractor.ts
@@ -1,6 +1,7 @@
 import * as path from 'path';
 
-import { ConfirmResult, DependencyInstaller, FileDownloader, InstallerSequence, InstallResult, Location, ProgressListener, Success, ZipExtractor } from '..';
+import { ConfirmResult, DependencyInstaller, FileDownloader, InstallerSequence, InstallResult, Success, ZipExtractor } from './';
+import { Location, ProgressListener } from '../util';
 
 /**
  * Extension of RemoteZipExtractor with the following features:
