+++ src/index.ts
@@ -1,2 +1,5 @@
+// exports all modules of this package
+
 export * from './dependencies';
 export * from './util';
+export * from './vscode-util';
