+++ src/util/index.ts
@@ -1,3 +1,5 @@
+// exports all modules in the current directory
+
 export * from './Location';
 export * from './Platform';
-export * from './Progress';
+export * from './ProgressListener';
