+++ src/vscode-util/index.ts
@@ -1,0 +1,3 @@
+// exports all modules in the current directory
+
+export * from './Progress';
